refactor: inline Dukascopy job form
