+++ frontend/src/routes/settings/index.stories.tsx
@@ -1,15 +1,15 @@
 import { RouterProvider, createMemoryRouter } from "react-router-dom";
 import type { Meta, StoryObj } from "@storybook/react";
 import { within, expect } from "storybook/test";
-import DukascopyJobSettings from "./index";
+import Settings from "./index";
 
-const router = createMemoryRouter([{ path: "/", element: <DukascopyJobSettings /> }]);
+const router = createMemoryRouter([{ path: "/", element: <Settings /> }]);
 
 const meta = {
-  component: DukascopyJobSettings,
+  component: Settings,
   render: () => <RouterProvider router={router} />,
   parameters: { layout: "fullscreen" },
-} satisfies Meta<typeof DukascopyJobSettings>;
+} satisfies Meta<typeof Settings>;
 
 export default meta;
 
@@ -18,6 +18,7 @@ type Story = StoryObj<typeof meta>;
 export const Default: Story = {
   play: async ({ canvasElement }) => {
     const canvas = within(canvasElement);
-    await expect(canvas.getByText("Dukascopyジョブ管理")).toBeInTheDocument();
+    await expect(canvas.getByText("設定")).toBeInTheDocument();
+    await expect(canvas.getByLabelText("通貨ペア")).toBeInTheDocument();
   },
 };
